Embtoolkit: Fisrt commit
This is the first commit in the embedded system toolkit git repository, formaly
embtoolkit. This porject aims to design a tool which helps on building an
embedded linux system. The goal of this porject is to use EGLIBC as C Library
for linux systems generated.

This initial commit is manely just a backup in order to save work done.

Signed-off-by: Abdoulaye Walsimou Gaye <walsimou@walsimou.com>