Toolchain: eglibc: add libm OPTION_EGLIBC_LIBM option group
This option group enables default configuration of the math
library.
Not selecting this option group removes most of the extended and
double precision math functions and replaces them with wrappers
to the single precision couterparts.
Doing so greatly degrades quality of calculations carried
out by the functions of the math library, but also significantly
reduces the size of the libm.
This option group is useful for systems that do not rely on
precise floating point math.

Signed-off-by: Abdoulaye Walsimou Gaye <walsimou@walsimou.com>